Move wpx back to its own file
