+++ src/runtime/index.js
@@ -1,7 +1,6 @@
 import registerDirectives from './directives';
 import registerComponents from './components';
 import { init } from './router';
-export { wpx } from './hooks';
 
 /**
  * Initialize the initial vDOM.
