+++ src/runtime/utils.js
@@ -18,20 +18,3 @@ export const createRootFragment = (parent, replaceNode) => {
 		},
 	});
 };
-
-// Deep Merge
-const isObject = (item) =>
-	item && typeof item === 'object' && !Array.isArray(item);
-
-export const deepMerge = (target, source) => {
-	if (isObject(target) && isObject(source)) {
-		for (const key in source) {
-			if (isObject(source[key])) {
-				if (!target[key]) Object.assign(target, { [key]: {} });
-				deepMerge(target[key], source[key]);
-			} else {
-				Object.assign(target, { [key]: source[key] });
-			}
-		}
-	}
-};
